requirements.txt
absl-py==2.1.0
altair==5.3.0
asgiref==3.8.1
asttokens 
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama 
comm 
contourpy==1.2.1
cycler==0.12.1
debugpy
decorator 
exceptiongroup
executing
filelock==3.14.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
gast==0.5.4
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
gunicorn==23.0.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata
intel-openmp==2021.4.0
ipykernel
ipython
itsdangerous==2.2.0
jedi 
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client 
jupyter_core 
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest_asyncio
networkx==3.3
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging
parso 
pickleshare 
pillow==10.3.0
platformdirs 
pluggy==1.2.0
prompt-toolkit 
protobuf==4.25.3
psutil
psycopg2-binary==2.9.9
pure-eval 
pyarrow==16.1.0
pydeck==0.9.1
Pygments
pyparsing==3.1.2
python-dateutil
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
setuptools==68.2.2
six 
smmap==5.0.1
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data 
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.5.22
toml==0.10.2
toolz==0.12.1
tornado
tqdm==4.66.4
traitlets
typer==0.9.0
typing_extensions
tzdata==2024.1
ultralytics==8.2.91
ultralytics-thop==2.0.6
urllib3==2.2.1
watchdog==4.0.1
wcwidth
Werkzeug==3.0.3
wheel==0.41.2
wrapt==1.16.0
WTForms==3.1.2
zipp
